+++ server/handlers.js
@@ -1,23 +1,26 @@
-const addUser = require("./handlers/addUser")
-const getUser = require("./handlers/getUser")
-const getUsers = require("./handlers/getUsers")
-const logIn = require("./handlers/logIn")
-const updateUser = require("./handlers/updateUser")
-const addList = require("./handlers/addList")
-const addReview = require("./handlers/addReview")
-const movieDetails = require("./handlers/movieDetails")
-const searchGenre = require("./handlers/searchGenre")
-const searchQuery = require("./handlers/searchQuery")
-const homeMovies = require("./handlers/homeMovies")
-const getGenres = require("./handlers/getGenres")
-const getReviews = require("./handlers/getReviews")
-const addFriend = require("./handlers/addFriend")
-const removeFriend = require("./handlers/removeFriend")
-const verifyToken = require("./handlers/verifyToken")
-const middleware = require("./handlers/middleware")
-const likeReview = require("./handlers/likeReview")
-const removeLikeReview = require("./handlers/removeLikeReview")
-const addComment = require("./handlers/addComment")
+// users handlers
+const getUsers = require("./handlers/users/getUsers")
+const getUser = require("./handlers/users/getUser")
+const addUser = require("./handlers/users/addUser")
+const logIn = require("./handlers/users/logIn")
+const updateUser = require("./handlers/users/updateUser")
+const addFriend = require("./handlers/users/addFriend")
+const removeFriend = require("./handlers/users/removeFriend")
+const verifyToken = require("./handlers/users/verifyToken")
+const middleware = require("./handlers/users/middleware")
+// interactions handlers
+const addList = require("./handlers/interactions/addList")
+const getReviews = require("./handlers/interactions/getReviews")
+const addReview = require("./handlers/interactions/addReview")
+const likeReview = require("./handlers/interactions/likeReview")
+const removeLikeReview = require("./handlers/interactions/removeLikeReview")
+const addComment = require("./handlers/interactions/addComment")
+// TMDB API handlers
+const movieDetails = require("./handlers/tmdb/movieDetails")
+const searchGenre = require("./handlers/tmdb/searchGenre")
+const searchQuery = require("./handlers/tmdb/searchQuery")
+const homeMovies = require("./handlers/tmdb/homeMovies")
+const getGenres = require("./handlers/tmdb/getGenres")
 
 module.exports = {
     addUser,
