+++ server/handlers/interactions/addList.js
