+++ server/handlers/interactions/likeReview.js
@@ -2,11 +2,11 @@ const {MongoClient, ObjectId} = require("mongodb")
 require("dotenv").config()
 const {MONGO_URI} = process.env
 
-const removeLikeReview = async(req, res) =>{
+const likeReview = async(req, res) =>{
     const {username, name, reviewId} = req.body
     if(!username || !reviewId || !name){
-        return res.status(404).json({
-            status: 404,
+        return res.status(400).json({
+            status: 400,
             message: `The request is not complete`
         })
     }
@@ -15,22 +15,22 @@ const removeLikeReview = async(req, res) =>{
         await client.connect()
         const db = client.db("movie")
         const id = new ObjectId(reviewId)
-        const removeLike = await db.collection("reviews").updateOne({_id: id}, {$pull:{likes: {username: username, name: name}}})
-        if(removeLike.matchedCount === 0){
+        const like = await db.collection("reviews").updateOne({_id: id}, {$push:{likes: {username: username, name: name}}})
+        if(like.matchedCount === 0){
             return res.status(404).json({
             status:404,
             message: `review id (${reviewId}) was not found`
             })
         }
-        if(removeLike.modifiedCount === 0){
+        if(like.modifiedCount === 0){
             return res.status(409).json({
             status: 409,
-            message: `${username} was not able to remove the like for the review`
+            message: `${username} was not able to like the review`
             })
         }
         res.status(200).json({
             status: 200,
-            message: "The review was successfully unliked"
+            message: "The review was successfully liked"
         })
     }
     catch(error){
@@ -44,4 +44,4 @@ const removeLikeReview = async(req, res) =>{
     }
 }
 
-module.exports = removeLikeReview
+module.exports = likeReview
