+++ server/handlers/tmdb/getGenres.js
