+++ server/handlers/tmdb/homeMovies.js
@@ -27,8 +27,8 @@ const homeMovies = async(req, res) => {
         })
     }
     catch(error){
-        res.status(500).json({
-            status:500,
+        res.status(502).json({
+            status:502,
             message: error.message
         })
     }
