+++ server/handlers/tmdb/movieDetails.js
@@ -38,8 +38,8 @@ const movieDetails = async(req, res) => {
         })
     }
     catch(error){
-        res.status(500).json({
-            status:500,
+        res.status(502).json({
+            status:502,
             message: error.message
         })
     }
