+++ server/handlers/users/addFriend.js
