+++ server/handlers/users/removeFriend.js
