+++ server/handlers/users/verifyToken.js
